Implement coin game object
In general, a game score system is now possible, since it's built to be flexible.

+++ src/Game/LevelSelector.h
@@ -170,5 +170,5 @@ private:
 	Float mLevelGuiAnim;
 
 	LS_CachedVariable<Float, Int> mTimer;
-	LS_CachedVariable<Int, Int> mCoins;
+	LS_CachedVariable<Float, Int> mCoins;
 };
